backend/requirements.txt
flask
flask-cors
python-docx
openpyxl
reportlab
python-pptx
PyPDF2
